refactor(lualine): Move lualine generation to util
